backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-app"
version = "0.1.0"
description = "A Flask web application integrating with Ollama."
dependencies = [
    "requests>=2.28.0",
    "ollama>=0.1.0",
    "python-dotenv>=1.1.1",
    "fastapi>=0.110.0",
    "uvicorn>=0.32.0",
]
[tool.hatch.build.targets.wheel]
packages = ["src"]